% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NOAA.R
\name{clean_cache}
\alias{clean_cache}
\title{Clean cache of NOAA data}
\usage{
clean_cache()
}
\value{
The path to \code{extdata} (invisibly).
}
\description{
Remove all cached NOAA data files from package's \code{extdata} directory.
}
\examples{
library(fs)

# get NOAA data with caching
get_NOAA("temperature", 1, "January", TRUE)

# path
rel_pt <- url_parser("temperature", 1, "January", TRUE)$local # relative path
abs_pt <- path_package("oceanexplorer", rel_pt) # absolute path

# check
file_exists(abs_pt)

# clean cache
clean_cache()

# check again
file_exists(abs_pt)
}
