% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-labeller.R
\name{env_parm_labeller}
\alias{env_parm_labeller}
\title{Parsing expressions for plot labels}
\usage{
env_parm_labeller(var, prefix = character(1), postfix = character(1))
}
\arguments{
\item{var}{Environmental parameter.}

\item{prefix}{Prefix.}

\item{postfix}{Postfix.}
}
\value{
Expression
}
\description{
Parsing expressions for plot labels
}
