% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_NOAA.R
\name{filter_NOAA}
\alias{filter_NOAA}
\title{Filter NOAA}
\usage{
filter_NOAA(NOAA, depth, coord = NULL, output = "point")
}
\arguments{
\item{NOAA}{Dataset of the WORLD OCEAN ATLAS.}

\item{depth}{Depth in meters}

\item{coord}{List with named elements: \code{lon} for longitude in degrees, and
\code{lat} for latitude in degrees.}

\item{output}{Geometry point (default \code{"point"}). Additional geometries are
currently not supported.}
}
\value{
Either a stars object or sf dataframe
}
\description{
Filter NOAA
}
\examples{
\dontrun{
# get atlas
NOAAatlas <- get_NOAA("oxygen", 1, "annual")
# filter atlas for specific depth an coordinates
filter_NOAA(NOAAatlas, 30)
}
}
