% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_module.R
\name{filter_ui}
\alias{filter_ui}
\alias{filter_server}
\title{NOAA filter module}
\usage{
filter_ui(id)

filter_server(id, NOAA)
}
\arguments{
\item{id}{Namespace id shiny module.}

\item{NOAA}{Reactive value of NOAA dataset.}
}
\value{
Shiny module.
}
\description{
NOAA filter module
}
