% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NOAA.R
\name{get_NOAA}
\alias{get_NOAA}
\title{Obtain NOAA WORLD OCEAN ATLAS dataset}
\usage{
get_NOAA(var, spat_res, av_period, cacheNOAA = TRUE)
}
\arguments{
\item{var}{The chemical or physical variable of interest.}

\item{spat_res}{Spatial resolution, either 1 or 5 degree grid-cells (numeric)
.}

\item{av_period}{Temporal resolution, either \code{"annual"}, specific seasons
(e.g. \code{"winter"}), or month (e.g. \code{"August"}).}

\item{cacheNOAA}{Caching the extracted files under \code{inst/extdata}.}
}
\value{
Star object.
}
\description{
The function can also automatically caches the extracted files.
}
\examples{
\dontrun{
get_NOAA("oxygen", 1, "annual")
}
}
