% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_module.R
\name{input_ui}
\alias{input_ui}
\alias{input_server}
\title{NOAA data module}
\usage{
input_ui(id)

input_server(id)
}
\arguments{
\item{id}{Namespace id shiny module.}
}
\value{
Shiny module.
}
\description{
NOAA data module
}
