% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NOAA.R
\name{list_NOAA}
\alias{list_NOAA}
\title{List cached NOAA data files}
\usage{
list_NOAA()
}
\value{
A character vector containing the names of the files in the specified
directories (empty if there were no files). If a path does not exist or is
not a directory or is unreadable it is skipped.
}
\description{
List all cached NOAA data files from package's \code{extdata} directory.
}
\examples{

# show cached NOAA files
list_NOAA()

}
