% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_addin.R
\name{NOAA_addin}
\alias{NOAA_addin}
\title{Ocean explorer addin}
\usage{
NOAA_addin(server = NOAA_server(extended = FALSE))
}
\arguments{
\item{server}{Server function.}

\item{extended}{Boolean whether to launch the extended app (default = \code{TRUE})
or the limited version for usage as a RStudio gadget.}
}
\value{
Rstudio gadget
}
\description{
Wrapper function that launches the NOAA RStudio addin
}
\examples{

\dontrun{
# run RStudio addin (can also be launched from `Addins` dropdown menu)
NOAA_addin()
}
}
