% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_app.R
\name{NOAA_app}
\alias{NOAA_app}
\alias{NOAA_server}
\title{Ocean explorer app}
\usage{
NOAA_app(server = NOAA_server())

NOAA_server(extended = TRUE)
}
\arguments{
\item{server}{Server function.}

\item{extended}{Boolean whether to launch the extended app (default = \code{TRUE})
or the limited version for usage as a RStudio gadget.}
}
\value{
Shiny app
}
\description{
Wrapper function that launches the NOAA app.
}
\examples{

\dontrun{
# run app
NOAA_app()
}
}
