% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_data.R
\docType{data}
\name{NOAA_data}
\alias{NOAA_data}
\title{NOAA variable names and units.}
\format{
A tibble with 7 rows and 3 variables:
\describe{
\item{variable}{oceanographic variable}
\item{unit}{variable unit}
\item{citation}{citation of the dataset}
}
}
\source{
\url{https://www.ncei.noaa.gov/products/world-ocean-atlas}
}
\usage{
NOAA_data
}
\description{
A dataset containing the variable names and units of data from NOAA made
available through this package.
}
\keyword{datasets}
