% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanexplorer-package.R
\docType{package}
\name{oceanexplorer-package}
\alias{oceanexplorer}
\alias{oceanexplorer-package}
\title{oceanexplorer: Explore our Planet's Oceans with NOAA}
\description{
This package enables easy exploration of the world ocean atlas of the US agency National Oceanic and Atmospheric Administration (NOAA). It includes functions to extract NetCDF data from the repository and code to visualise several physical and chemical parameters of the ocean. A Shiny app further allows interactive exploration of the data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://utrechtuniversity.github.io/oceanexplorer/ https://utrecht-university.shinyapps.io/oceanexplorer/}
}

}
\author{
\strong{Maintainer}: Martin Schobben \email{schobbenmartin@gmail.com} (\href{https://orcid.org/0000-0001-8560-0037}{ORCID})

}
\keyword{internal}
