% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NOAA.R
\name{plot_NOAA}
\alias{plot_NOAA}
\title{Plotting the global NOAA WORLD OCEAN ATLAS}
\usage{
plot_NOAA(NOAA, points = NULL, epsg = NULL, limit = NULL)
}
\arguments{
\item{NOAA}{Dataset of the WORLD OCEAN ATLAS.}

\item{points}{Add locations of extracted point geometry.}
}
\value{
Ggplot
}
\description{
Plotting the global NOAA WORLD OCEAN ATLAS
}
\examples{
\dontrun{
# data
NOAA <- get_NOAA("oxygen", 1, "annual")

# base (surface depth)
base <- filter_NOAA(NOAA,  0)

# coordinates
points <- filter_NOAA(NOAA, 1, list(lon = c(-160, -120), lat =  c(11,12)))

#plot
plot_NOAA(base, points)
}
}
