% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_module.R
\name{plot_ui}
\alias{plot_ui}
\alias{plot_server}
\title{NOAA plot module}
\usage{
plot_ui(id)

plot_server(id, NOAA)
}
\arguments{
\item{id}{Namespace id shiny module.}

\item{NOAA}{Reactive value of NOAA dataset.}
}
\value{
Shiny module.
}
\description{
NOAA plot module
}
