% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{env_parm_labeller}
\alias{env_parm_labeller}
\title{Parsing expressions for plot labels}
\usage{
env_parm_labeller(var)
}
\arguments{
\item{var}{Environmental parameter.}
}
\value{
Expression
}
\description{
Parsing expressions for plot labels
}
