% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_module.R
\name{filter_ui}
\alias{filter_ui}
\alias{filter_server}
\title{NOAA filter module}
\usage{
filter_ui(id, citation, extended = TRUE)

filter_server(id, NOAA, variable, external, extended = TRUE)
}
\arguments{
\item{id}{Namespace id shiny module.}

\item{citation}{Additional space for citation element.}

\item{extended}{Boolean whether to build the extended module
(default = \code{TRUE}).}

\item{NOAA}{Reactive value of NOAA dataset.}

\item{variable}{Reactive value for the selected variable name.}

\item{external}{Reactive value for filter operation based on plot selection.}
}
\value{
Shiny module.
}
\description{
NOAA filter module
}
