% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_module.R
\name{filter_ui}
\alias{filter_ui}
\alias{filter_server}
\title{NOAA filter module}
\usage{
filter_ui(id, extended = TRUE)

filter_server(
  id,
  NOAA,
  external,
  ivars = c("depth", "lon", "lat"),
  extended = TRUE
)
}
\arguments{
\item{id}{Namespace id shiny module.}

\item{extended}{Boolean whether to build the extended module
(default = \code{TRUE}).}

\item{NOAA}{Reactive value of NOAA dataset.}

\item{external}{Reactive values for latitude, longitude and depth from plot
module.}

\item{ivars}{Character vector for the variables for filtering.}
}
\value{
Shiny module.
}
\description{
NOAA filter module
}
