% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NOAA.R
\name{get_NOAA}
\alias{get_NOAA}
\title{Obtain NOAA WORLD OCEAN ATLAS dataset}
\usage{
get_NOAA(var, spat_res, temp_res, year)
}
\arguments{
\item{var}{The chemical or physical variable of interest.}

\item{spat_res}{Spatial resolution, either 1 or 5 degree grid-cells (numeric)
.}

\item{temp_res}{Temporal resolution, either \code{"annual"}, \code{"monthly"} and
\code{"daily"}.}

\item{year}{Year of publication (numeric).}
}
\value{

}
\description{
Obtain NOAA WORLD OCEAN ATLAS dataset
}
\examples{
\dontrun{
get_NOAA("oxygen", 1, "annual", 2001)
}
}
