% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NOAA.R
\name{get_NOAA}
\alias{get_NOAA}
\title{Obtain NOAA WORLD OCEAN ATLAS dataset}
\usage{
get_NOAA(var, spat_res, av_period)
}
\arguments{
\item{var}{The chemical or physical variable of interest.}

\item{spat_res}{Spatial resolution, either 1 or 5 degree grid-cells (numeric)
.}

\item{av_period}{Temporal resolution, either \code{"annual"}, specific seasons
(e.g. \code{"winter"}), or month (e.g. \code{"August"}).}
}
\value{
Star object.
}
\description{
Obtain NOAA WORLD OCEAN ATLAS dataset
}
\examples{
\dontrun{
get_NOAA("oxygen", 1, "annual")
}
}
