% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_module.R
\name{input_ui}
\alias{input_ui}
\alias{citation_ui}
\alias{output_ui}
\alias{input_server}
\alias{output_server}
\title{NOAA data module}
\usage{
input_ui(id, citation = NULL, extended = TRUE)

citation_ui(id)

output_ui(id)

input_server(id)

output_server(id, locations, variable)
}
\arguments{
\item{id}{Namespace id shiny module.}

\item{extended}{Boolean whether to build the extended module
(default = \code{TRUE}).}

\item{locations}{Reactive value for the dataset containing the locations
coordinates.}

\item{variable}{Reactive value for the selected variable name.}
}
\value{
Shiny module.
}
\description{
NOAA data module
}
