% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_addin.R
\name{NOAA_addin}
\alias{NOAA_addin}
\title{Ocean explorer addin}
\usage{
NOAA_addin(cache = FALSE)
}
\arguments{
\item{cache}{Caching the extracted NOAA file in the package's \code{extdata}
directory (default = \code{FALSE}). Size of individual files is around 12 Mb. Use
\code{\link[=list_NOAA]{list_NOAA()}} to list cached data resources.}
}
\value{
Rstudio gadget
}
\description{
Wrapper function that launches the NOAA RStudio addin
}
\examples{

if (curl::has_internet() && interactive()) {

# run RStudio addin (can also be launched from `Addins` dropdown menu)
NOAA_addin()

}
}
