% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_app.R
\name{NOAA_app}
\alias{NOAA_app}
\alias{NOAA_server}
\title{Ocean explorer app}
\usage{
NOAA_app(cache = FALSE)

NOAA_server(extended = TRUE, cache)
}
\arguments{
\item{cache}{Caching the extracted NOAA file in the package's \code{extdata}
directory (default = \code{FALSE}). Size of individual files is around 12 Mb. Use
\code{\link[=clean_cache]{clean_cache()}} to remove all cached data.}

\item{extended}{Boolean whether to launch the extended app (default = \code{TRUE})
or the limited version for usage as a RStudio gadget.}
}
\value{
Shiny app
}
\description{
Wrapper function that launches the NOAA app.
}
\examples{

\dontrun{
# run app
NOAA_app()
}
}
