% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_app.R
\name{NOAA_app}
\alias{NOAA_app}
\alias{NOAA_server}
\title{Ocean explorer app}
\usage{
NOAA_app(cache = FALSE)

NOAA_server(extended = TRUE, cache)
}
\arguments{
\item{cache}{Caching the extracted NOAA file in the package's \code{extdata}
directory (default = \code{FALSE}). Size of individual files is around 12 Mb. Use
\code{\link[=list_NOAA]{list_NOAA()}} to list cached data resources.}

\item{extended}{Boolean whether to build the extended module
(default = \code{TRUE}).}
}
\value{
Shiny app
}
\description{
Wrapper function that launches the NOAA app.
}
\examples{

if (curl::has_internet() && interactive()) {

# run app
NOAA_app()

}
}
