% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanexplorer-package.R
\docType{package}
\name{oceanexplorer-package}
\alias{oceanexplorer}
\alias{oceanexplorer-package}
\title{oceanexplorer}
\description{
This package allows easy exploration of the world ocean atlas of
the US agency National Oceanic and Atmospheric Administration (NOAA).
It includes functions to extract NetCDF data from the repository and code to
visualize several physical and chemical parameters of the ocean. A Shiny app
further allows interactive exploration of the data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://martinschobben.github.io/oceanexplorer/}
  \item \url{https://martinschobben.shinyapps.io/oceanexplorer/}
  \item \url{https://www.ncei.noaa.gov/products/world-ocean-atlas}
  \item \url{https://github.com/MartinSchobben/oceanexplorer}
  \item Report bugs at \url{https://github.com/MartinSchobben/oceanexplorer/issues}
}

}
\author{
\strong{Maintainer}: Martin Schobben \email{schobbenmartin@gmail.com} (\href{https://orcid.org/0000-0001-8560-0037}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Peter Bijl (\href{https://orcid.org/0000-0002-1710-4012}{ORCID}) [contributor]
}

}
\keyword{internal}
