% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NOAA.R
\name{plot_NOAA}
\alias{plot_NOAA}
\title{Plotting the global NOAA WORLD OCEAN ATLAS}
\usage{
plot_NOAA(NOAA)
}
\arguments{
\item{NOAA}{Dataset of the WORLD OCEAN ATLAS.}

\item{depth}{Depth to filter.}
}
\value{
Ggplot
}
\description{
Plotting the global NOAA WORLD OCEAN ATLAS
}
\examples{
\dontrun{
get_NOAA("oxygen", 1, "annual", 2001) \%>\% plot_NOAA()
}
}
