% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NOAA.R
\name{plot_NOAA}
\alias{plot_NOAA}
\title{Plotting the global NOAA WORLD OCEAN ATLAS}
\usage{
plot_NOAA(
  NOAA,
  depth = NULL,
  points = NULL,
  epsg = NULL,
  limit = NULL,
  rng = NULL
)
}
\arguments{
\item{NOAA}{Dataset of the WORLD OCEAN ATLAS.}

\item{depth}{Depth in meters.}

\item{points}{Add locations of extracted point geometry.}

\item{epsg}{The epsg used to project the data (currently supported 4326, 3031
and 3995).}

\item{limit}{The limits of the axis.}

\item{rng}{A vector of two numeric values for the range of the environmental
parameter.}
}
\value{
Ggplot
}
\description{
Plotting the global NOAA WORLD OCEAN ATLAS
}
\examples{
\dontrun{
# data
base <- get_NOAA("oxygen", 1, "annual")

# plot
plot_NOAA(base)

# coordinates
points <- filter_NOAA(NOAA, 1, list(lon = c(-160, -120), lat =  c(11,12)))

# plot
plot_NOAA(base, points)
}
}
