% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_module.R
\name{plot_ui}
\alias{plot_ui}
\alias{plot_server}
\title{NOAA plot module}
\usage{
plot_ui(id)

plot_server(id, NOAA, points, back, reset, depth_slider)
}
\arguments{
\item{id}{Namespace id shiny module.}

\item{NOAA}{Reactive value of NOAA dataset.}

\item{points}{Add locations of extracted point geometry.}

\item{back}{Reactive value for back button.}

\item{reset}{Reactive value for reset button.}
}
\value{
Shiny module.
}
\description{
NOAA plot module
}
