% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reproject}
\alias{reproject}
\alias{reproject.sf}
\alias{reproject.stars}
\title{Reprojecting spatial objects to new epsg}
\usage{
reproject(obj, epsg, ...)

\method{reproject}{sf}(obj, epsg)

\method{reproject}{stars}(obj, epsg)
}
\arguments{
\item{var}{Environmental parameter.}
}
\value{
sf or stars object
}
\description{
Reprojecting spatial objects to new epsg
}
