% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reproject}
\alias{reproject}
\alias{reproject_}
\alias{reproject_.sf}
\alias{reproject_.stars}
\title{Reprojecting spatial objects to new epsg}
\usage{
reproject(obj, epsg = NULL, ...)

reproject_(obj, epsg = NULL, ...)

\method{reproject_}{sf}(obj, epsg)

\method{reproject_}{stars}(obj, epsg)
}
\arguments{
\item{obj}{The sf or stars object to be reprojected.}

\item{epsg}{The projection.}

\item{...}{Currently not supported.}
}
\value{
sf or stars object
}
\description{
Reprojecting spatial objects to new epsg
}
