% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_module.R
\name{table_ui}
\alias{table_ui}
\alias{table_server}
\title{NOAA table module}
\usage{
table_ui(id, download = NULL, extended = TRUE)

table_server(id, NOAA, back, reset, extended = TRUE)
}
\arguments{
\item{id}{Namespace id shiny module.}

\item{download}{Add download button.}

\item{extended}{Boolean whether to build the extended module
(default = \code{TRUE}).}

\item{NOAA}{Reactive value of NOAA dataset.}

\item{back}{Reactive value for back button.}

\item{reset}{Reactive value for reset button.}
}
\value{
Shiny module.
}
\description{
NOAA table module
}
